// find-your-job/back/src/dto/register.dto.ts
export interface RegisterDto {
  name: string;
  email: string;
  password: string;
  confirmPassword: string;
  role?: "candidate" | "recruiter";
  company?: string;
  category?: string;
  country: string;
  state: string;
  city: string;
  phone?: string;
  about?: string;
  termsAccepted: boolean;
}